+++ project/crud-microservice-app/golang/testing-project/crudOperations.go
@@ -12,7 +12,6 @@ import (
 	"strings"
 	"time"
 
-	"go.mongodb.org/mongo-driver/bson/primitive"
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
 	"go.mongodb.org/mongo-driver/mongo/readpref"
@@ -1154,161 +1153,6 @@ func giveAllLearnr(w http.ResponseWriter, req *http.Request) {
 	}
 	theReturnMessage.ReturnedLearnRNames = learnrNameMap //Add our final OrgMap
 
-	//Format the JSON map for returning our results
-	theJSONMessage, err := json.Marshal(theReturnMessage)
-	//Send the response back
-	if err != nil {
-		errIs := "Error formatting JSON for return in giveAllLearnROrgs: " + err.Error()
-		logWriter(errIs)
-	}
-	fmt.Fprint(w, string(theJSONMessage))
-}
-
-/* This function takes parameters for LearnR sorting and returns an array */
-func specialLearnRGive(w http.ResponseWriter, req *http.Request) {
-	canCrud := true
-	//Declare data to return
-	type ReturnMessage struct {
-		TheErr          []string `json:"TheErr"`
-		ResultMsg       []string `json:"ResultMsg"`
-		SuccOrFail      int      `json:"SuccOrFail"`
-		ReturnedLearnrs []Learnr `json:"ReturnedLearnrs"`
-	}
-	theReturnMessage := ReturnMessage{}
-	theReturnMessage.SuccOrFail = 0 //Initially set to success
-
-	//Unwrap from JSON
-	bs, err := ioutil.ReadAll(req.Body)
-	if err != nil {
-		theErr := "Error reading the request from specialLearnRGive: " + err.Error() + "\n" + string(bs)
-		theReturnMessage.ResultMsg = append(theReturnMessage.ResultMsg, theErr)
-		theReturnMessage.TheErr = append(theReturnMessage.TheErr, theErr)
-		theReturnMessage.SuccOrFail = 1
-		logWriter(theErr)
-		fmt.Println(theErr)
-		canCrud = false
-	}
-
-	//Decalre JSON we recieve
-	type TheSpecialCases struct {
-		CaseSearch       []int  `json:"CaseSearch"`
-		OrganizationName string `json:"OrganizationName"`
-		Tag              string `json:"Tag"`
-		LearnRName       string `json:"LearnRName"`
-		EntryAmountFrom  int    `json:"EntryAmountFrom"`
-		EntryAmountTo    int    `json:"EntryAmountTo"`
-	}
-
-	//Marshal it into our type
-	var theitem TheSpecialCases
-	json.Unmarshal(bs, &theitem)
-
-	fmt.Printf("DEBUG: Here is our item: %v\n", theitem)
-
-	//Do CRUD operations if allowed
-	if canCrud {
-		/* Begin building crud operation based on our criteria */
-		collection := mongoClient.Database("learnR").Collection("learnr") //Here's our collection
-		var fullConditions bson.M = make(bson.M)
-		findOptions := options.Find()
-		theFind := 0 //A counter to track how many Learnrs we find
-
-		/* Alter the filter/findoptions based on criteria */
-		if theitem.CaseSearch[0] == 0 {
-			//Do nothing, just get all Learnrs
-		}
-		//Add tags into our search
-		if theitem.CaseSearch[1] == 0 {
-			fullConditions["tags"] = "cool"
-		}
-		//Add a Name into our search
-		if theitem.CaseSearch[2] == 0 {
-			fullConditions["name"] = bson.M{"$regex": primitive.Regex{
-				Pattern: "[" + theitem.LearnRName + "]",
-			}}
-		}
-		//fmt.Printf("DEBUG: Here is our fullConditions: %v\n", fullConditions)
-		/*DEBUG: Add cases later for more criteria */
-		/* Run the mongo query after fixed filter/findoptions */
-		find, err := collection.Find(theContext, fullConditions, findOptions)
-		if err != nil {
-			panic("Error getting this query...: " + err.Error())
-		} else if find.Err() != nil || err != nil {
-			if strings.Contains(err.Error(), "no documents in result") {
-				returnedErr := "No documents returned; may be that there are no Learnrs yet or search was bad..."
-				fmt.Println(returnedErr)
-				logWriter(returnedErr)
-				theReturnMessage.SuccOrFail = 0
-				theReturnMessage.ResultMsg = append(theReturnMessage.ResultMsg, returnedErr)
-				theReturnMessage.TheErr = append(theReturnMessage.TheErr, returnedErr)
-				theReturnMessage.ReturnedLearnrs = []Learnr{}
-			} else {
-				returnedErr := "There was a Mongo Error: " + err.Error()
-				fmt.Println(returnedErr)
-				logWriter(returnedErr)
-				theReturnMessage.SuccOrFail = 1
-				theReturnMessage.ResultMsg = append(theReturnMessage.ResultMsg, returnedErr)
-				theReturnMessage.TheErr = append(theReturnMessage.TheErr, returnedErr)
-				theReturnMessage.ReturnedLearnrs = []Learnr{}
-			}
-		} else {
-			//Found Learnr, decode to return
-			for find.Next(theContext) {
-				var theLearnRReturned Learnr
-				err := find.Decode(&theLearnRReturned)
-				if err != nil {
-					returnedErr := "For " +
-						", there was an error decoding document from Mongo: " + err.Error()
-					fmt.Println(returnedErr)
-					logWriter(returnedErr)
-					theReturnMessage.SuccOrFail = 1
-					theReturnMessage.ResultMsg = append(theReturnMessage.ResultMsg, returnedErr)
-					theReturnMessage.TheErr = append(theReturnMessage.TheErr, returnedErr)
-					theReturnMessage.ReturnedLearnrs = []Learnr{}
-				} else if len(theLearnRReturned.Name) <= 1 {
-					returnedErr := "For " +
-						", there was an no document from Mongo: " + err.Error()
-					fmt.Println(returnedErr)
-					logWriter(returnedErr)
-					theReturnMessage.SuccOrFail = 1
-					theReturnMessage.ResultMsg = append(theReturnMessage.ResultMsg, returnedErr)
-					theReturnMessage.TheErr = append(theReturnMessage.TheErr, returnedErr)
-					theReturnMessage.ReturnedLearnrs = []Learnr{}
-				} else {
-					//Successful decode, add this to our array
-					theReturnMessage.ReturnedLearnrs = append(theReturnMessage.ReturnedLearnrs, theLearnRReturned)
-				}
-				theFind = theFind + 1
-			}
-			find.Close(theContext)
-		}
-		//Declare results, see if we have errors
-		if theReturnMessage.SuccOrFail >= 1 {
-			theErr := "There are a number of errors for returning these special Learnrs..."
-			theReturnMessage.ResultMsg = append(theReturnMessage.ResultMsg, theErr)
-			theReturnMessage.TheErr = append(theReturnMessage.TheErr, theErr)
-		} else if len(theReturnMessage.ReturnedLearnrs) <= 0 {
-			theErr := "No learnrs returned...this could be the site's first deployment with no Learnrs or a faulty search!"
-			theReturnMessage.ResultMsg = append(theReturnMessage.ResultMsg, theErr)
-			theReturnMessage.TheErr = append(theReturnMessage.TheErr, theErr)
-			theReturnMessage.SuccOrFail = 0
-		} else {
-			theErr := "No issues returning Learnrs"
-			theReturnMessage.ResultMsg = append(theReturnMessage.ResultMsg, theErr)
-			theReturnMessage.TheErr = append(theReturnMessage.TheErr, theErr)
-			theReturnMessage.SuccOrFail = 0
-		}
-	} else {
-		//Error, return an error back and log it
-		returnedErr := "Had an issue getting special Learnrs"
-		logWriter(returnedErr)
-		theReturnMessage.SuccOrFail = 1
-		theReturnMessage.ResultMsg = append(theReturnMessage.ResultMsg, returnedErr)
-		theReturnMessage.TheErr = append(theReturnMessage.TheErr, returnedErr)
-		theReturnMessage.ReturnedLearnrs = []Learnr{}
-	}
-
-	//fmt.Printf("DEBUG: Here is what we are going to send back: %v\n", theReturnMessage.ReturnedLearnrs)
 	//Format the JSON map for returning our results
 	theJSONMessage, err := json.Marshal(theReturnMessage)
 	//Send the response back
